[tests] Add TestsFlavor parameter to RenameTestCases task
After adding more logging in
https://github.com/xamarin/xamarin-android/commit/a6eb47aea, it turned
out that the path to the logcat file is wrong. Note the directory of
the logcat in the following log message:

    Could not find file `/Users/builder/jenkins/workspace/xamarin-android-pr-pipeline-release/xamarin-android/build-tools/scripts/../../bin/TestRelease-Profiled-Aot/logcat-Release-Profiled-Aot-Mono.Android_Tests.txt`.  Will not be including `adb logcat` output.

The `bin/TestRelease-Profiled-Aot/` part should be `bin/TestRelease/`
instead.

This patch changes it so, that we pass `Configuration` and
`TestsFlavor` separately to the task and the logcat path should now be
correct.