[tests] Add RenameTestCases.TestsFlavor property (#2894)
Commit a6eb47aea stated:

> This "feels like" there's a mismatch in directories being used to
> store the `logcat-*` files, which is why they're not found.

It turned out that the path to the logcat file *is* wrong.  Note the
directory of the logcat file in the following log message:

	Could not find file `…/xamarin-android/build-tools/scripts/../../bin/TestRelease-Profiled-Aot/logcat-Release-Profiled-Aot-Mono.Android_Tests.txt`.  Will not be including `adb logcat` output.

The `bin/TestRelease-Profiled-Aot` directory should instead be
`bin/TestRelease`, and the directory contains `-Profiled-Aot` because
the provided `RenameTestCases.Configuration` value contained
`$(TestsFlavor)`, which was `-Profiled-Aot` in this case.

Improve the `<RenameTestCases/>` task so that the
`RenameTestCases.Configuration` contains *just* the
`$(Configuration)` value, and split out the `$(TestsFlavor)` value
into a separate `RenameTestCases.TestsFlavor` property.  This should
allow `RenameTestCases.GetTestCaseInfo()` to construct an appropriate
filesystem path for the `adb logcat` output.

[tests] Add RenameTestCases.TestsFlavor property (#2894)