[Xamarin.Android.Build.Tasks] Support Filtering Assemblies
As part of the AndroidX migration system we need to
be able to skip over the older support library
assemblies for certain build steps.

The following metadata is supported on the `ReferencePath`
and `ReferenceDependencyPaths` ItemGroups.

	AndroidSkipAddToPackage
	AndroidSkipResourceProcessing
	AndroidSkipJavaStubGeneration
	AndroidSkipResourceExtraction

`AndroidSkipResourceProcessing` is the renamed
`SkipAndroidResourceProcessing` metadata item.
This is used to skip case conversion on those
resources.

`AndroidSkipResourceExtraction` will stop the
resources in that library being extracted into
the `IntermediateOutputPath`.

`AndroidSkipJavaStubGeneration` when `true` this
item will stop java code being generated for the
library in quetion.

`AndroidSkipAddToPackage` is used by `BuildApk`
task to skip over the library and NOT add it to
the final APK.

To use these new metadata items you can define your
own `ItemGroup` as follows

	<ItemGroup>
		<_MyStuff Include="Xamarin.Android.Support.v4">
			<AndroidSkipAddToPackage>True</AndroidSkipAddToPackage>
		</_MyStuff>
	</ItemGroup>

Then add this new `ItemGroup` to the `AndroidCustomMetaDataForReferences`
like so.

	<ItemGroup>
		<AndroidCustomMetaDataForReferences Include="@(_MyStuff)" />
	</ItemGroup>

The `AndroidCustomMetaDataForReferences` ItemGroup will then
be processed by the `_AddAndroidCustomMetaData` target. This
will add the new metadata to the `ReferencePath`
and `ReferenceDependencyPaths` ItemGroups.