Use correct `strip` path when building with mono archives
Mono Android build still creates a standalone NDK toolchain as part of its
provisioning step and places it in `~/android-toolchain/toolchains` and its
subdirectories. Xamarin.Android then uses this location to find the `strip`
utility which we call after the build is done. However, when Mono archives are
used and the Mono build does *not* take place, the standalone toolchain is never
created (as Xamarin.Android itself doesn't use it) and thus we fail to find the
required `strip` binaries and the build fails

This commit fixes the problem by always using strip from the prebuilt toolchains
as shipped with the NDK.