[Xamarin.Android.Build.Tasks] Clean up redundant files.
Commit 121b81aa added the new `runtime_java.jar`.
What it did not take into account was what happens
when a user upgrades but does not do a clean build.

What happens is we have old redundant files left
in the `$(IntermediateOutputPath)android\src` and
`$(IntermediateOutputPath)android\bin\classes`.

This then causes errors such as

	error: duplicate class: mono.MonoPackageManager_Resources

when compiling the java code.

So what we need to do is clean up after ourselves. We do
this by checking to see if the following file exists

	$(MonoAndroidIntermediate)android\src\mono\app\ApplicationRegistration.java

If that file is there, then we need to clean up all the
other known files. We can not do a clean all the time since
we still generate `MonoPackageManager_Resources`. The
contents have changed though. So if the above file is in
place, we can safely remove all the others.