[build-tools] Increase the emulator disk space. (#2927)
We have more and more tests running on the emulator now.  Sometimes
we run out of space to run the tests.

Increases the disk space.

We also handle a problem with the `<CreateAndroidEmulator/>` error
processing.  `advmanager` may emit the following warning.

	Warning: cvc-pattern-valid: Value '' is not facet-valid with respect to pattern '[a-zA-Z0-9_-]+' for type 'idType'.
		org.xml.sax.SAXParseException; lineNumber: 134; columnNumber: 341; cvc-pattern-valid: Value '' is not facet-valid with respect to pattern '[a-zA-Z0-9_-]+' for type 'idType'.
		at com.sun.org.apache.xerces.internal.util.ErrorHandlerWrapper.createSAXParseException(ErrorHandlerWrapper.java:203)
		at com.sun.org.apache.xerces.internal.util.ErrorHandlerWrapper.error(ErrorHandlerWrapper.java:134)

Only the first line was being picked up as a warning, the rest was
being reported as an error.  The result was the emulator never ran!