[Xamarin.Android.Build.Tasks] don't copy pdb/mdb to $(OutputPath)
Looking at the build output of a simple app:

    > ls .\samples\HelloWorld\bin\Debug\*.pdb
      Length Name
      ------ ----
        1024 HelloLibrary.pdb
        1304 HelloWorld.pdb
    13943324 Mono.Android.pdb
      153048 Mono.Security.pdb
     1545052 mscorlib.pdb
      106160 System.ComponentModel.Composition.pdb
      393532 System.Core.pdb
       43076 System.Net.Http.pdb
      795468 System.pdb
      325732 System.Runtime.Serialization.pdb
       77008 System.ServiceModel.Internals.pdb
      917588 System.Xml.pdb

There are *not* accompanying `.dll` files for many of these symbols.
Some of these symbols are also quite big!

I don't think we need to be copying these files to `$(OutputPath)` at
all; it doesn't make sense if their accompanying assembly isn't there!

If we remove the `<CopyIfChanged/>` calls in our targets, MSBuild will
handle copying the appropriate symbol files that are needed.

After this change, the directory looks like:

    > ls .\samples\HelloWorld\bin\Debug\*.pdb
      Length Name
      ------ ----
      1024 HelloLibrary.pdb
      1304 HelloWorld.pdb

~~ Results ~~

I did a performance comparison using the Xamarin.Forms app in this
repo. It seems the extra `<CopyIfChanged/>` call was taking between
25-50ms, but the bigger savings here is to not copy a 13MB
Mono.Android.pdb file. There could be bigger savings for developers
that have slower hardware (or don't have SSDs).

Overall it seems this change saves up to 50ms for a small project on
initial build, or incremental builds where an assembly changed.