[build] Make MinGW build on mac generate less noise
For some reason the Windows cross-build on macOS using MinGW leaves behind a lot
of temporary LTO (Link-Time Optimization) files, the result of passing the
`-flto` flag to the compilers. The temporary files (`*.ltrans*`), around 280 of
them, are then included in our vsix packages, increasing their size while being
completely useless for the end user.

This commit fixes the issue by turning off LTO for MinGW targets. The
optimization is not really useful there and such fix is faster than tracking why
MinGW on macOS (installed from Homebrew) is behaving the way it is.