[Xamarin.Android.Build.Tasks] <FilterAssemblies/> support missing TFI (#2958)
Context: https://www.nuget.org/packages/Refractored.Controls.CircleImageView/
Context: https://github.com/Azure-Samples/MyDriving

The MyDriving sample app currently fails to build on master with:

	Resources/layout/fragment_profile.axml(2): error APT0000: attribute civ_border_width (aka com.microsoft.mydriving:civ_border_width) not found.

The failure happens with both `aapt` and `aapt2`.

This layout is using a custom view such as:

	<?xml version="1.0" encoding="utf-8"?>
	<ScrollView
	    xmlns:android="http://schemas.android.com/apk/res/android"
	    xmlns:local="http://schemas.android.com/apk/res-auto">
	    ...
	    <refractored.controls.CircleImageView local:civ_border_width="0dp" />
	    ...
	</ScrollView>

This comes from the `Refractored.Controls.CircleImageView` NuGet
package.

In 5ec3e3a, I added a `<FilterAssemblies/>` MSBuild task that appears
to be to blame. It is not returning
`Refractored.Controls.CircleImageView.dll`, but it needs to!

[`Refractored.Controls.CircleImageView.dll`][0] has no
`[assembly: System.Runtime.Versioning.TargetFrameworkAttribute]`
(via `ildasm`):

	// C:\src\des\MyDriving\packages\Refractored.Controls.CircleImageView.1.0.1\lib\MonoAndroid10\Refractored.Controls.CircleImageView.dll
	// Refractored.Controls.CircleImageView, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null
	// Global type: <Module>
	// Architecture: x86
	// Runtime: v4.0.30319
	// Hash algorithm: SHA1
	
	using Android.Runtime;
	using System.Reflection;
	using System.Runtime.CompilerServices;
	using System.Security;
	using System.Security.Permissions;
	
	[assembly: AssemblyTitle("Refractored.Controls.CircleImageView")]
	[assembly: AssemblyDescription("")]
	[assembly: AssemblyConfiguration("")]
	[assembly: AssemblyCompany("")]
	[assembly: AssemblyProduct("")]
	[assembly: AssemblyCopyright("2015 Refractored LLC/James Montemagno")]
	[assembly: AssemblyTrademark("")]
	[assembly: NamespaceMapping(Java = "de.hdodenhof.circleimageview", Managed = "Refractored.Controls")]
	[assembly: RuntimeCompatibility(WrapNonExceptionThrows = true)]
	[assembly: SecurityPermission(8, SkipVerification = true)]
	[assembly: AssemblyVersion("1.0.0.0")]
	[module: UnverifiableCode]

It is indeed a `MonoAndroid` assembly, since it references
`Mono.Android.dll`.  It is weird, though...


~~ What should we do? ~~

One idea is to look for `Mono.Android.dll` references in each
assembly *instead* of `TargetFrameworkIdentifier`.

Looking at the assemblies in Xamarin.Forms, there is a complication to
this idea:

  * `Xamarin.Forms.Core.dll` (a NetStandard library) references
      * `Xamarin.Forms.Platform.dll` (a MonoAndroid library?) references
          * `Xamarin.Forms.Platform.Android.dll` (A MonoAndroid library)

`Xamarin.Forms.Platform.dll` has
`[assembly:TargetFramework ("MonoAndroid, ...")]`, but does not have
an assembly reference to `Mono.Android.dll`.

Should we "recursively" looked for references to `Mono.Android.dll`
and "bubbled up" that information?  If we did so, then
`Xamarin.Forms.Core.dll` would be treated as a MonoAndroid assembly.
This in turn is undesirable, both because it means *all* libraries
which reference would in turn need the extra processing done by
`<GenerateJavaStubs/>`/etc. and because *we know* that this extra
processing is unnecessary: *only* a reference to `Xamarin.Forms.Core`
cannot result in the generate of Java code.


~~ Conclusion ~~

I think we should stick with `TargetFrameworkIdentifier`, and in the
rare case it is missing look for a `Mono.Android.dll` reference for
the single assembly.  This seems like it is going to cover all cases
to me, and we still will get good performance.

So we will cover:

  * `Xamarin.Forms.Platform.dll` counted as a "MonoAndroid" assembly
    since it has a `TargetFrameworkIdentifier` (but no reference).
  * `Refractored.Controls.CircleImageView.dll` counted as a
    "MonoAndroid" assembly. It has no `TargetFrameworkIdentifier, but
    has a reference to `Mono.Android.dll`.

Changes include:

  * `<FilterAssemblies/>` needs to also check for an assembly
    reference to `Mono.Android` as a fallback.
  * `<ResolveAssemblies/>` now adds `%(HasMonoAndroidReference)=True`
    item metadata.
  * When creating the `@(_ResolvedUserMonoAndroidAssemblies)` item
    group, we also check for `%(HasMonoAndroidReference)=True`.

I added a few tests to verify these scenarios.

[0]: https://github.com/jamesmontemagno/CircleImageView-Xamarin.Android