[Xamarin.Android.Build.Tasks] cache aapt2 version in AppDomain
If you have `$(AndroidUseAapt2)` set to `True` (which is also the
default on master), the `<ResolveAndroidTooling/>` tooling task seems
to take longer:

    aapt2:
        83 ms  ResolveAndroidTooling                      2 calls
    aapt:
         4 ms  ResolveAndroidTooling                      2 calls

I figured out this was due to how we shell out to `aapt2 version` to
check the version of `aapt2`.

We should use `RegisterTaskObject` and cache the result of this call!

* I used a key that contains the path to `aapt2`, so if that
* I don't think we actually need to invalidate the cached value on
  anything else because we are shipping `aapt2` now.

But what if there is a Xamarin.Android update?

If you install a new VSIX, VS on Windows makes you restart the IDE. If
you install a PKG on Mac, you need to restart the IDE *anyway*,
because it keeps all system MSBuild `.targets` files in memory...
Weird things will happen if you install a Xamarin.Android PKG on Mac
and don't restart VS for Mac.

After this change, when the `aapt2 version` call is cached:

    2 ms  ResolveAndroidTooling                      2 calls

This was the "Hello World" sample project in this repo.