[tests] Measure activity displayed time (#2974)
Add special-case handled measurement of activity displayed message
from Android `ActivityManager`.

We have to special case code for that, because our normal message
measurements come from the app process, while the activity manager is
a separate process and thus would not be matched by our timing
definitions, which use the PID to pre-filter the matches.

The activity manager adds message to the log when an activity is
displayed.  We hope that this correlates with actual time when the
activity is displayed on screen and interactive with the user, and
thus is interesting time to measure.

The activity messages look like this:

	04-05 15:50:00.421  1201  1239 I ActivityManager: Displayed com.google.android.deskclock/com.android.deskclock.alarms.AlarmActivity: +136ms
	04-05 16:00:01.157  1201  1239 I ActivityManager: Displayed com.google.android.deskclock/com.android.deskclock.alarms.AlarmActivity: +145ms
	04-05 16:36:23.803  1201  1239 I ActivityManager: Displayed Xamarin.Forms_Performance_Integration/xamarin.forms.performance.integration.MainActivity: +1s255ms
	04-05 16:38:14.092  1201  1239 I ActivityManager: Displayed org.androworks.meteorgram/.MeteorgramActivity: +580ms
	04-05 16:38:42.140  1201  1239 I ActivityManager: Displayed org.mozilla.firefox/org.mozilla.gecko.BrowserApp: +469ms
	04-05 16:38:54.098  1201  1239 I ActivityManager: Displayed com.whatsapp/.HomeActivity: +545ms

Example measurement for our XF test:

	last-Debug,JNI.init-Debug,init-Debug,OnCreateBegin-Debug,OnCreateEnd-Debug,OnStartBegin-Debug,OnStartEnd-Debug,OnResumeBegin-Debug,OnResumeEnd-Debug,ActivityDisplayed-Debug
	1908,492,492,585,1332,1337,1338,1338,1341,2047

*NOTE*: as can be seen above, the activity `Displayed` message
(2047ms in column 10; the last column) comes *long* after
`OnResumeEnd()` was called (1341ms column 9, 2nd-to-last column).