[Xamarin.Android.Build.Tests] Add support for testing the UI (#2855)
Context: https://github.com/xamarin/xamarin-android/issues/2608

Bumps to mono/debugger-libs/master@b45303.
Bumps to icsharpcode/nrefactory/master@0607a4.

Add support for testing that the app launches on device *and* can be
interacted with.  It does this by making use of some Android shell
features and commands

`adb shell input` allows us to send taps and clicks to the device.
It simulates user input.  This can be used to e.g. click a button.

`adb exec-out uiautomator` allows obtaining the complete UI layout
for the currently active Activity.  For example, to dump the current
Activity UI layout to the console:

	adb exec-out uiautomator dump /dev/tty

This returns a XML:

	<hierarchy rotation="0">...

We can parse this XML looking for any ui element we want.  We can
then get the bounds of the control and pass that information to
`adb shell input tap`.

Since we are now able to read the displayed UI, we can also add some
`TimeZoneInfo` tests, to ensure that when the timezome is changed:

	adb shell su root setprop persist.sys.timezone "America/New_York"

the UI is correctly updated to display the appropriate timezone name.

Finally, add tests to check that the debugger -- via
mono/debugger-libs -- can actually connect to the Application.

The debugger-related tests only run if the commercial bits of the
Xamarin.Android SDK are available.

As all of these new tests *require* that an emulator or device be
available, they have been split out into a new
`tests/MSBuildDeviceIntegration` project.  This is built as part of
the main solution.  However, it is run as part of the
`make run-apk-tests` target.  This is so we can be sure the emulator
is running.

In the future any MSBuild related tests that require a device should
be placed in this new project.